Agrego chequeo de que no elimine una cancha con reservas

+++ backend/src/main/java/ar/uba/fi/ingsoft1/todo_template/field/FieldService.java
@@ -78,6 +78,10 @@ public class FieldService {
             throw new EntityNotFoundException("Solo el propietario del campo puede eliminarlo");
         }
 
+        if (!field.getReservations().isEmpty()) {
+            throw new IllegalArgumentException("No se puede eliminar un campo con reservas asociadas");
+        }
+
         fieldRepository.delete(field);
     }
